Add compatibility tests and rewrite byte parsing
The byte parsing using `Array#pack` and `String#unpack` was not working
the way I thought it was, since I was only using the maximum value as
a test. That was a mistake that I discovered by adding the compatibility
tests.

Writing these bye conversions in Ruby will mean that they will run
slower. I'm curious how much slower, so we should write some benchmarks.
With benchmarks, we should be able to decide whether it's fast enough or
should be rewritten for performance.