Add support for ActiveRecord (#7)
You can now easily define KSUID fields within ActiveRecord. These fields
can be made to store either the string or binary representation of the
KSUID in the database. There is also a mixin that you can use to make
sure you're handling a KSUID object instead of the primitive version
stored in the database.

As part of this, there is now documentation in the project README to
describe how to use the ActiveRecord integration. There was also some
cleanup and refactoring that was done during the work.

Lastly, we're now testing against more recent Rubies and have
integration tests with all version of Rails 5.