Fix all possible mutation failures
There are a few mutation failures that are causing problems around
ActiveRecord. I cannot figure out how to filter them out, so we're left
with ~99.7% mutation coverage.

Note that this requires a patch that I submitted upstream that has been
merged, but not released as a new version yet. As such, we have that
commit locked in the Gemfile.

There is also a weird method removal within the Base62 tests that
removes an addition from ActiveSupport. This library should not rely on
ActiveSupport, so we cannot make that change to fail the mutation ... so
we undefine the method and let that handle the problem.