Benchmark and adjust the Base62 charset check
I had a hunch that the way that we were checking for character presence
in the Base62 class wasn't the most performant since it created an
iterator with a block. Thus, this benchmark: it checks the four
different ways that I could think of to perform this task.

Surprisingly, that was the second fastest way of doing it!

Using `Regexp#match?` is a little more than twice as fast as the current
implementation, so I opted to make the switch. Included below are some
timings on my laptop.

```
Warming up --------------------------------------
            include?    28.062k i/100ms
       Regexp#match?    68.160k i/100ms
             Array#-    10.556k i/100ms
          two finger     8.203k i/100ms
Calculating -------------------------------------
            include?    282.838k (± 0.6%) i/s -      1.431M in   5.060186s
       Regexp#match?    673.904k (± 0.9%) i/s -      3.408M in   5.057563s
             Array#-    104.379k (± 0.8%) i/s -    527.800k in   5.056886s
          two finger     79.321k (± 1.1%) i/s -    401.947k in   5.068036s

Comparison:
       Regexp#match?:   673903.6 i/s
            include?:   282838.5 i/s - 2.38x  (± 0.00) slower
             Array#-:   104379.4 i/s - 6.46x  (± 0.00) slower
          two finger:    79320.6 i/s - 8.50x  (± 0.00) slower
```