Fix a leaky test
When this test ran early on in the suite, it poisoned the rest of the
suite with the no-op generator. This led to nondeterministically failing
tests when you create more than one model with a unique KSUID column.

This change ensures that we reset the configuration to the default
random generator.