Fix handling of binary primary keys on JRuby
Due to an error in implementing the `ActiveRecord::Type#type` method for
both types, JRuby was loading an incorrect value when deserializing from
the database for binary types. This didn't affect the string type
because the default representation is a string.

This only affected JRuby, I think, because of how JRuby interfaces with
the Java internals. Since MRI doesn't have that translation layer, we
didn't see the same type of problem there.

I am backporting this change from the extraction of the ActiveRecord
behavior into a separate gem.