Update supported Rubies and Rails
This constrains support to currently supported Rubies. I know that JRuby
9.2 is still supported, but it targets Ruby 2.5 semantics, which is
out-of-date. Since versions prior to v0.4.0 had broken ActiveRecord
support on JRuby, I wonder if it's worth having in CI. I like to be
inclusive though, so I'm keeping JRuby 9.3.

Every time I need to do this on a project, I need to figure out the
compatibility matrix. [This][1] is helpful in doing so and so is the
[JRuby security document][2].

[1]: https://www.fastruby.io/blog/ruby/rails/versions/compatibility-table.html
[2]: https://github.com/jruby/jruby/blob/master/SECURITY.md