Deprecate KSUID::ActiveRecord
In preparation for v1.0.0, this change deprecates the
`KSUID::ActiveRecord` constant in favor of the newly introduced
`ActiveRecord::KSUID` that does the same thing.

For v1.0.0 we want to extract the ActiveRecord-related behavior into a
separate gem so that we aren't mixing concerns. The `ksuid` gem will be
only the datatype whereas the ActiveRecord integration will live in the
ActiveRecord gem. Any further integrations will receive their own gems.