Add support for prefixed KSUIDs
In some contexts, it can be helpful to prefix your KSUIDs with a tag to
identify what model it belongs to. For example, you might identify Event
KSUIDs with an `evt_` prefix and Customer KSUIDs with a `cus_` prefix.

To do so, you can now use the `KSUID.prefixed` method. You can also use
a prefixed KSUID within ActiveRecord. These interoperate with normal
KSUIDs in that they can convert to each other and order themselves
appropriately by time.