Allow non-Rails tests gracefully skip
One annoyance that I have when I come back to working on this gem after
a while is that I forget that I need to use Appraisals to run the whole
test suite.

This change prints a message instead of bombing out when running RSpec
outside of Appraisals.

Of course, this goes away once we split the gem ...