Allow disabling the automatic generation of KSUIDs
For certain cases, such as foreign keys, automatically generating a
value for the field is incorrect. In these cases, you can now pass
`auto_gen: false` to disable the automatic generation.

The number of options on the module builder is starting to seem like a
smell. Perhaps it's time to unwind it in favor of the plain attribute
API. We will retain it for backwards compatibility though.