Test activerecord-ksuid on multiple databases
The ActiveRecord database adapters all behave slightly differently,
particularly when it comes to binary data. In order to ensure that the
gem works on the most common permutations of database, Ruby version, and
Rails version, this change introduces testing on MySQL and PostgreSQL,
in addition to SQLite, during CI.

To aid in running tests against all these permutations, we now ship a
`docker-compose.yml` file for spinning up the databases we test against.
You can run this with either Podman Compose or Docker Compose.