Fix binary KSUIDs on PostgreSQL
PostgreSQL serializes binary data differently than SQLite, such that we
need to do an extra level of munging when deserializing data from the
database. I'm not sure if this is due to something within the `pg` gem,
`libpsql`, or Ruby, but filtering out an extra hexidecimal indicator on
the front end of the string fixes the issue and causes no extra issues
with MySQL or SQLite.