Fix local testing to work for JRuby and MRI
In getting JRuby to work on CI with MySQL, I accidentally changed too
many features of the test setup. When using JRuby with MySQL, you must
specify `mysql` as the driver, but when using MRI you need to use
`mysql2`.

This change handles that when using the `rake spec` helper, which is the
way we test locally.