Fix generator behavior for ksuid column types
This change makes it so `rails generate model example:ksuid` works as
expected. I am unsure how fragile this test is. Ideally, we would have a
test that runs the generator against a dummy app and asserts on the
outcome, but I don't have the spoons to manage that right now.